add viewport
